requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
xgboost==2.0.0
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.25.2
imbalanced-learn==0.11.0
shap==0.42.1
faker==19.3.1
PyYAML==6.0.1
streamlit==1.26.0
matplotlib==3.7.2
